+++ packages/react/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./VibeGrid";
 export * from "./VibeGridPlaceholder";
+export * from "./grid-density";